requirements.txt
# Core dependencies for Jira Search Mirror
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
flask>=2.3.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0